+++ config/config.exs
@@ -328,6 +328,7 @@ config :pleroma, :activitypub,
   unfollow_blocked: true,
   outgoing_blocks: true,
   follow_handshake_timeout: 500,
+  note_replies_output_limit: 5,
   sign_object_fetches: true,
   authorized_fetch_mode: false
 
@@ -483,6 +484,7 @@ config :pleroma, Oban,
     transmogrifier: 20,
     scheduled_activities: 10,
     background: 5,
+    remote_fetcher: 2,
     attachments_cleanup: 5,
     new_users_digest: 1
   ],